# tools/openapi.yaml
openapi: 3.0.3
info:
  version: 0.9.6
  title: LANraragi API
...
paths:
...
  /archives/upload:
    put:
      security:
        - api_key: []
      operationId: uploadArchive
      x-mojo-to: api-archive#create_archive
      summary: 🔑 Upload Archive
      description: >-
        Upload an Archive to the server.  

        If a SHA1 checksum of the Archive is included, the server will perform
        an optional in-transit, file integrity validation, and reject the upload
        if the server-side checksum does not match.
      tags:
        - archives
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Archive file to upload
                file_checksum:
                  type: string
                  description: SHA1 checksum of the archive for in-transit validation.
                  pattern: ^[a-fA-F0-9]{40}$
                category_id:
                  type: string
                  description: Category ID you'd want the archive to be added to.
                  minLength: 14
                  maxLength: 14
                tags:
                  type: string
                  description: >-
                    Set of tags you want to insert in the database alongside the
                    archive.
                title:
                  type: string
                  description: Title of the Archive.
                summary:
                  type: string
                  description: Summary of the Archive.
      responses:
        '200':
          description: Archive uploaded successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - operation
                  - success
                  - id
                properties:
                  operation:
                    type: string
                    description: Name of operation
                    example: upload
                  success:
                    type: integer
                    description: Returns 1 if operation was successful, else 0
                    enum:
                      - 1
                  id:
                    type: string
                    minLength: 40
                    maxLength: 40
                    description: ID of the uploaded archive
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '409':
          description: Duplicate archive
          content:
            application/json:
              schema:
                type: object
                required:
                  - operation
                  - success
                  - error
                  - id
                properties:
                  operation:
                    type: string
                    description: Name of operation
                    enum: 
                      - upload
                  success:
                    type: integer
                    enum:
                      - 0
                    description: Returns 0 for failed operation
                  error:
                    type: string
                    description: Error message indicating the archive already exists
                  id:
                    type: string
                    minLength: 40
                    maxLength: 40
                    description: ID of the existing archive  
                example:
                  operation: upload
                  error: This file already exists in the Library. Enable replace duplicated archive in config to replace old ones.
                  success: 0
                  id: 0b91b546850e881034833c73375558928ddcec7c
        '415':
          description: Unsupported file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '417':
          description: Checksum mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /archives/{id}:
    get:
      operationId: getArchive
      x-mojo-to: api-archive#serve_metadata
      deprecated: true
      summary: Get archive metadata (deprecated)
      description: Get Metadata (title, tags) for a given Archive.
      tags:
        - archives
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Archive to process.
          schema:
            type: string
            minLength: 40
            maxLength: 40
      responses:
        '200':
          description: Archive metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveMetadataJson'
        '400':
          description: No archive ID response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
    delete:
      security:
        - api_key: []
      operationId: deleteArchive
      x-mojo-to: api-archive#delete_archive
      summary: 🔑 Delete archive
      description: |-
        Delete both the archive metadata and the file stored on the server.  
        🙏 Please ask your user for confirmation before invoking this endpoint.
      tags:
        - archives
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Archive to process.
          schema:
            type: string
            minLength: 40
            maxLength: 40
      responses:
        '200':
          description: Archive deleted response
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    description: Name of the operation
                    enum:
                      - delete_archive
                  id:
                    type: string
                    minLength: 40
                    maxLength: 40
                    description: ID of the archive
                  filename:
                    type: string
                    description: Filename of the deleted archive
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                    description: Returns 1 if operation was successful, else 0
                example:
                  operation: delete_archive
                  success: 1
                  id: 75d18ce470dc99f83dc355bdad66319d1f33c82b
                  filename: big_chungus.zip
        '400':
          description: No archive ID response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /archives/{id}/metadata:
    get:
      operationId: getArchiveMetadata
      x-mojo-to: api-archive#serve_metadata
      summary: Get archive metadata
      description: Get Metadata (title, tags) for a given Archive. 
      tags:
        - archives
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Archive to process.
          schema:
            type: string
            minLength: 40
            maxLength: 40
      responses:
        '200':
          description: Archive metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveMetadataJson'
        '400':
          description: No archive ID response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
    put:
      security:
        - api_key: []
      operationId: updateArchiveMetadata
      x-mojo-to: api-archive#update_metadata
      summary: 🔑 Update archive metadata
      description: >-
        Update tags and title for the given Archive.  

        Data supplied to the server through this method will **overwrite** the previous data.
      tags:
        - archives
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Archive to process.
          schema:
            type: string
            minLength: 40
            maxLength: 40
        - name: title
          in: query
          required: false
          description: New Title of the Archive.
          schema:
            type: string
        - name: tags
          in: query
          required: false
          description: New Tags of the Archive.
          schema:
            type: string
        - name: summary
          in: query
          required: false
          description: New Summary of the Archive.
          schema:
            type: string
      responses:
        '200':
          description: Archive metadata updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
              example:
                operation: update_metadata
                success: 1
        '400':
          description: No archive ID response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
...
  /archives/{id}/categories:
    get:
      operationId: getArchiveCategories
      x-mojo-to: api-archive#get_categories
      summary: Get archive categories
      description: Get all the Categories which currently refer to this Archive ID.
      tags:
        - archives
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Archive to process.
          schema:
            type: string
            minLength: 40
            maxLength: 40
      responses:
        '200':
          description: Archive categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    description: Array of category metadata
                    items:
                      $ref: '#/components/schemas/CategoryMetadataJson'
                  operation:
                    type: string
                    description: Name of the operation
                    enum:
                      - find_arc_categories
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                    description: Returns 1 if operation was successful, else 0
                example:
                  operation: find_arc_categories
                  success: 1
                  categories:
                    - id: SET_1613080290
                      name: My great category
                      pinned: 0
                      search: null
                      archives:
                        - 0d50858d727723d856d2ab78564bb8e906e65f14
                        - 7f03b8b1f337e1e42b2c2890533c0de7479d41ca
                    - id: SET_1613080294
                      name: My other category
                      pinned: 1
                      search: null
                      archives:
                        - 0d50858d727723d856d2ab78564bb8e906e65f14
                        - 75d18ce470dc99f83dc355bdad66319d1f33c82b
        '400':
          description: No archive ID response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
...
  /archives/{id}/download:
    get:
      operationId: downloadArchive
      x-mojo-to: api-archive#serve_file
      summary: Download archive
      description: Download an Archive from the server.
      tags:
        - archives
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Archive to download.
          schema:
            type: string
            minLength: 40
            maxLength: 40
      responses:
        '200':
          description: Archive file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: No archive ID response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
              example:
                operation: download_archive
                error: No archive ID specified.
                success: 0
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header
      description: 'Use Authorization: Bearer <base64(api_key)>'
  schemas:
...
    OperationResponse:
      type: object
      required:
        - operation
        - success
      properties:
        operation:
          type: string
          description: Name of operation
        error:
          type: string
          description: Error message if any
        successMessage:
          type: string
          description: Success message if any
        success:
          type: integer
          description: Returns 1 if operation was successful, else 0
          enum:
            - 0
            - 1
...
    CategoryMetadataJson:
      type: object
      description: >-
        Category metadata.
      required:
        - id
      properties:
        archives:
          type: array
          description: Array of archive IDs (if this is a static category, empty otherwise)
          items:
            type: string
            minLength: 40
            maxLength: 40
        id:
          type: string
          description: ID of the category
          minLength: 14
          maxLength: 14
        name:
          type: string
          description: Name of the category
        pinned:
          oneOf:
            - type: integer
            - type: string
          description: Whether this category should be pinned in the UI
          enum:
            - 0
            - 1
        search:
          type: string
          nullable: true
          description: Category search filter (if this is a dynamic category, empty otherwise)
      example:
        id: SET_1613080290
        name: My great category
        pinned: 0
        search: null
        archives:
          - 0d50858d727723d856d2ab78564bb8e906e65f14
          - 7f03b8b1f337e1e42b2c2890533c0de7479d41ca
...
servers:
  - url: https://lrr.tvc-16.science/api